+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Accept `req` as argument to `parse`
+
 0.0.0 / 2014-06-13
 ==================